\name{AdaptFitOS-internal}
\alias{Predict.matrix.ospline.smooth}
\alias{Predict.matrix.tlspline.smooth}
\alias{smooth.construct.os.smooth.spec}
\alias{smooth.construct.tl.smooth.spec}
\alias{Predict.matrix}
\alias{Predict.matrix.lme}
\alias{aspmOS}
\alias{aspmFormReadOS}
\alias{aspmPenReadOS}
\alias{aspmKrigeRead}
\alias{aspGetTheta}
\alias{almeAux}
\alias{acompute.CTRinvC}
\alias{bdiag}
\alias{airls.ridge}
\alias{arg.searchOS}
\alias{aspGetThetaOS}
\alias{bs2}
\alias{drvbasis}
\alias{print.asp}
\alias{spmDesignOS}
\alias{aspOS}
\alias{spmOS}
\alias{scbTest}
\alias{plot.scbTest}
\alias{spmArgReadOS}
\alias{aspmKrigeReadOS}

\title{Internal AdaptFitOS objects}
\description{Internal functions}
\details{These are not to be called by the user}
\keyword{internal}